Fix rebase issues, update dependencies, and add aot cache generation

+++ shared/src/main/java/com/faforever/neroxis/map/placement/SpawnPlacer.java
@@ -88,9 +88,9 @@ public class SpawnPlacer {
                      .subtract(
                              new BooleanMask(map.getSize() + 1, random.nextLong(), spawnMask.getSymmetrySettings())
                                      .drawSymmetryLines(spawnMask.getSymmetrySettings().teamSymmetry())
-                                     .inflate((float) teamSeparation / spawnMask.getSymmetrySettings()
-                                                                                .spawnSymmetry()
-                                                                                .getNumSymPoints()))
+                                     .inflate(teamSeparation / spawnMask.getSymmetrySettings()
+                                                                        .spawnSymmetry()
+                                                                        .getNumSymPoints()))
                      .fillEdge(map.getSize() / 32, false)
                      .limitToSymmetryRegion();
         Vector2 location = spawnMaskCopy.getRandomPosition();
